backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Polygon API
polygon-api-client==1.12.4

# Data processing
numpy==1.26.2
pandas==2.1.3

# Redis caching (optional)
redis==5.0.1

# Async HTTP
httpx==0.25.2
aiohttp==3.9.1

# Supabase
supabase==2.3.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment variables
python-dotenv==1.0.0